requirements.txt
#python = 3.10
torch~=2.4.0
torchaudio
torchvision
-f https://download.pytorch.org/whl/cu121
hydra-core==1.3.2
tqdm==4.65.2
imageio
pysodmetrics

ultralytics

colorama
Augmentor
scikit-learn

opencv-python>=4.5.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.4.0
Pillow>=8.3.0
psutil>=5.8.0
thop>=0.1.0